backend/restAPI/requirements.txt
emoji==2.11.0
numpy==1.23.0
protobuf==3.20.0
nltk==3.7
stanza==1.8.1
pymongo==4.7.1
fastapi==0.89.1
python-dotenv==1.0.1
uvicorn==0.18.0
deeppavlov==1.6.0
wiki-ru-wordnet==1.0.3